requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
gunicorn
pandas
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage
google-cloud-storage
pymysql
SQLAlchemy